requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
google-generativeai==0.3.0
sentence-transformers==2.2.2
numpy==1.24.3
beautifulsoup4==4.12.2
PyPDF2==3.0.1
requests==2.31.0
python-multipart==0.0.6
python-docx==1.1.0